virtio-net: cap objcache pagesizes at PAGESIZE_2M
Commit 3eb8533b introduced a bug whereby large virtio-net queue
sizes (e.g. 4096 on gce) would result in an objcache pagesize being greater
than the maximum guaranteed alignment of the parent (linear backed) heap. This
could result in an assertion failure in the objcache code when allocating a
parent page that is not aligned to the objcache pagesize.

While the id heap aligns allocations to the rounded power-of-2 of the
allocation size (a property that ought to be parameterized...), the physical
id heap ranges themselves are only guaranteed to be aligned to 2M. Thus,
allocations from the physical id heap or linear backed heap may also only be
aligned up to a size of 2M.

The problem is avoided in this case by capping the objcache pagesize at
PAGESIZE_2M when setting up the virtio-net caches; a larger pagesize is
unnecessary anyhow, as the objcache meta accounts for less than .1% of a
PAGESIZE_2M.