Unix resolve_dir(): fix handling of non-directory dirfd values
Filesystem-related syscalls that take a directory file descriptor
and a file path should return -ENOTDIR if the supplied file
descriptor is an open descriptor but does not correspond to a
directory. The current code is assuming that if the file descriptor
can be resolved it refers to an open directory; this may cause
access to invalid memory addresses if the above assumption does not
hold.
This change adds a check for the file descriptor type in the
resolve_dir() macro, which returns -ENOTDIR if the check fails.
A few runtime tests have been amended with test cases that would
cause the kernel to misbehave (e.g. try to access a filesystem
struct through an invalid pointer) without this fix.