TFS: fix directory cleanup when writing directory entries to log
When a TFS metadata tuple is written to the TFS log, if it refers
to a directory entry it should be cleaned up (recursively if the
entry is a directory itself) by temporarily removing references to
parent directories, otherwise the log will contain useless entries.
This is not being done in the current code by the tfs_rename()
function, which is not doing the cleanup recursively, with the
result that renaming a non-empty directory produces spurious TFS
log entries.
In addition, if a TFS log rebuild is triggered during the first
call to log_write_eav() by filesystem_write_eav(), the
filesystem_log_rebuild() function fixes up the entire filesystem
tree, which may cause the second call to log_write_eav() to do a
log write of tuples with references to their parents.

This change fixes the above two issues by adding the logic to clean
up and fix up a directory tree to the filesystem_write_eav()
function, which now takes an additional boolean argument indicating
whether a cleanup (and subsequent fixup) should be performed. The
rename runtime test is being amended with a new test case that
would produce spurious TFS log entries without this fix.