stage2, uefi: allocate untyped buffers from tagged region
This addresses a long-standing bug in targets that use a rewind tag for
values (specifically the stage2 and uefi bootloaders), whereby an untyped
buffer allocation was being allocated without creating space for the tag
type. This could manifest in an incorrect tag type being read for a buffer. To
remedy this, an additional tagged region heap is created of type tag_unknown,
and untyped buffer values (currently only coming from decode_value() when
using an old TFS encoding for backward compatibility) are allocated from that
heap.