getrandom(2): properly handle faulting user memory
If an unresolvable page fault occurs when random data is being
written to a user-supplied buffer in a getrandom() syscall, the
syscall should return -EFAULT (or the number of bytes written if
the fault occurs during `getrandom_deferred`). Since the buffer is
accessed while holding the chacha20 mutex, a
random_buffer_aborted() function has been added in order to release
the mutex after a fault.
This change fixes a kernel crash occurring if a machine runs out of
memory when a getrandom() syscall is invoked with a user buffer
that needs to be faulted in.