Page cache: improve handling of I/O requests under memory pressure
A file read or write request from userspace requires the kernel to
allocate pagecache pages to cover the file range being read or
written. If the system is under memory pressure, the kernel may be
unable to allocate such pages, and the existing code deals with
this occurrence by returning an error to userspace.
This change set improves handling of out-of-memory conditions
during file read/write syscalls so that no error is returned to
userspace if memory can be reclaimed in order to satisfy (at least
partially) a read/write request. A boolean `flush` parameter has
been added to the mm_service() function, which can now flush cached
data to disk (suspending the current context until the flush
operation is complete) if the desired amount of memory could not be
reclaimed by the registered memory cleaners. In the pagecache code,
mm_service() is now called when servicing an I/O request under
memory pressure, thereby making the relevant syscall block if
necessary.
If a write request cannot be fully processed even after flushing
cached data to disk, the subset of pages that could be processed
are written immediately, while the remaining pages are processed
subsequently (possibly after reclaiming the memory occupied by
previously processed pages).
If a read request cannot be fully processed even after flushing
cached data to disk, the subset of pages that could be processed
are copied to the user buffer, and the relevant syscall returns
the length of data that has been copied (it is not an error if
fewer bytes than requested are read during a syscall invocation).

In addition, a few other changes are being made in order to make
the kernel more robust under memory pressure:
- the amount of reserved memory that cannot be touched by a
pagecache page allocation has been increased from 4 to 8 MB, so
that when a page cannot be allocated there is still a large enough
amount of free memory that can be used e.g. to flush pagecache
contents to disk; the amount of reserved memory is not being
changed for "low-memory" instances, so that the pagecache is still
usable even if the total amount of physical memory is low
- the objcache used to allocate pagecache completions has now the
locked heap as parent, instead of the reserve heap wrapper, because
allocation of pagecache completions is critical when doing
operations such as syncing the pagecache to disk, thus the kernel
should be able to allocate a completion even after a page
allocation failed

The "unable to allocate page" error message in the objcache code has
been changed to a warning message, because after these changes
failure to allocate from an objcache can happen multiple times
without resulting in a fatal error condition.