demand_anonymous_page(): improve handling of out-of-memory errors
demand_anonymous_page() is called when the user application
accesses a memory address for which a page of physical memory has
not been allocated yet. If the system is under memory pressure and
a page cannot be allocated, the existing code sends a SIGBUS signal
to the user program, even if there is reclaimable memory that could
be used to allocate the requested page.
This change improves handling of out-of-memory conditions so that
cached data are flushed to disk (suspending the context that
requested a new page) when a page allocation fails: this allows
reclaiming memory that can then be used to allocate the requested
page.
In new_zeroed_pages(), the "cannot get physical page" error message
is being changed to a debug message, because after these changes
failure to allocate a page from physical memory can happen without
resulting in a user-visible error condition.