fault in stack pages
Up to now, Nanos would allocate and map an unnecessarily physically-contiguous
2MB chunk of memory for the userspace stack. This could be wasteful for
low-memory configurations both in terms of allocated but potentially unused
stack space as well as memory fragmentation created by the large, contiguous
allocation.

With this change, only a single page (PROCESS_STACK_PREALLOC_SIZE) is
allocated for stack usage at first, allowing further stack pages to be faulted
in as needed. A new vmap flag, VMAP_FLAG_STACK, has been added to identify the
stack mapping, a fault on which is handled in the same way as an anonymous
mmap.