Pagecache: put a limit on the size of scatter-gather lists
Excessively large I/O requests from userspace may lead to large SG
lists being set up in the kernel, which may cause memory allocation
problems. PR #1752 modified the pagecache committing process by
introducing a limit on the size of each write request. However,
since #1816, TFS extents created when reserving storage space for
files being written are set as "uninited", which means that if a
write request does not cover the entire range of an extent, the TFS
driver initializes to zero the range that is not covered. As a
result, when the user program writes to a file in large chunks, a
large part of newly created TFS extents needs to be set to zero
when committing the file data to disk, which is inefficient
(because the zeroed area is going to be overwritten shortly with
"real" data), and defeats the purpose of splitting large ranges
into smaller ones (in fact, it makes things worse, because a call
to zero_blocks() needs to allocate one SG buffer for each page in
the range being zeroed).
This change increases the SG list size limit in the pagecache
committing process, and introduces a corresponding limit (32 MB) on
the extent size in the TFS driver. This prevents in typical use
cases the inefficient and potentially memory-hungry zeroing of
parts of uninited TFS extents. In addition, the SG list size limit
is enforced on the SG lists created by the pagecache code when
reading from a file, and the code in fs/fs.c to read an entire file
has been amended in order to deal with partial reads.