pagecache_volume: remove write_error status field
In the existing code, the write_error status field is set (and then
never cleared) on a pagecache volume when an error occurs during a
commit of any node beloning to the volume; then, this error is
reported at any subsequent write requests on any node belonging to
the volume. This causes write errors due to temporary failures
(e.g. shortage of available physical memory during a pagecache
flush to disk) to prevent any further file writes from the user
program, even if the error occurred on a pagecache node different
from the node being modified by the program.
This change removes the write_error status field from the pagecache
volume struct. Since file writes are flushed to disk
asynchronously, the user program cannot rely on detecting write
errors via the return code of write syscalls, and the only way to
ensure that file data is physically stored on the underlying disk
is to use fsync().