fault in user program pages
Until this point, user program executables have been loaded and mapped, in
their entirety, using physically-contiguous backed allocations. This practice
has the effect of creating pressure on the memory system in the form of
fragmentation, which is especially problematic in low memory configurations.

With this change, the user program file and program loader, if any, are now
demand paged. Rather than reading the program with filesystem_read_entire(), a
small portion is read, which is enough to contain the ELF program headers. The
program headers are parsed using load_elf(), but instead of mapping loaded
data directly into user memory, the elf_map_handlers instead add pagecache
node information to the vmaps covering program text and data, allowing pages
for these areas to be faulted in. BSS mappings are flagged as such and treated
as anonymous mappings (like stack and heap mappings).

To support this change, load_elf() now passes only file offsets (instead of
buffer physical addresses) to elf_map_handlers as the second argument. The
handlers must find the buffer physical address by enclosing the buffer start
and finding the physical address based on the given offset. The logic for
handling non-page-aligned starts to bss sections has been moved from
load_elf() to the handlers themselves, and the bss portion of any loadable
segment is expressed with a new 'bss_size' argument, rather than handled in
two separate invocations of the elf_map_handler.

Note that this feature is disabled (in favor of a statically-mapped load in
entirety) when ltrace is in use, because the ltrace code needs to be able to
make instruction-level modifications to the program in memory. It may also be
disabled with the 'static_map_program' config option.

exec_elf() now takes a path string for the program file and performs the file
loading itself. If 'ltrace' or 'static_map_program' are specified, the whole
program binary is loaded at once and kept resident in memory. Otherwise, only
a minimum necessary to read the program headers (ELF_PROGRAM_LOAD_MIN_SIZE, or
one page) is loaded at first. Additional ELF data may be subsequently read as
necessary, for instance to access the path string for the program loader, or
to read in symbol data if 'ingest_program_symbols' is specified. When running
in the fault-in mode, buffered file data is released after ELF header parsing
has completed.