Improved OOM handling in page faults
If a page fault for a user program fails, a SIGBUS signal is generated and
control is returned to the user program. If the page fault failed for an out-
of-memory condition and the program has a signal handler, it's possible
to hit another OOM page fault but the new signal is ignored since it hasn't
been handled yet. This can create a loop in the user program where no
progress is made because of the OOM condition, but neither does the
program exit.

This change adds a check for a OOM page fault that will kill the program with
SIGKILL if enough OOM page faults have occured in a short amount of time.
A successful page fault or a long enough time since the last OOM page fault
will reset the check. This behavior is modeled after the Linux OOM killer.

In order to distinguish OOM page faults from other kinds of errors,
do_demand_page and its callees had to be modified to return a status
which can be used to distinguish the failure type.