Allow multiple entries in expected_error_code
Since different signal types can be generated from the same root cause,
such as an out of memory condition, it is useful to allow a test to accept
more than one kind of error code. This change preserves the original string
type in the expected_error_code, but now also checks to see if it is a vector
and iterates and compares all the codes in the list.