Add additional e2e tests
This change expands the e2e tests beyond testing just for web servers. The
new e2e code can run an arbitrary program, and the new testing code checks
for the exit status message at the end of the output to determine a test
success or failure. The AsyncCmdStart code has been modified to use a
CommandContext with a custom Cancel function in order to provide a
reliable timeout and process status for the command. An additional
Postrun step is added in order to handle any cleanup required after a
test is run.

The new tests include several versions of large memory allocations with
several different language packages, a test for disk full conditions and
data corruption, and a test of some of the cloud init functionality.