Add support for building with clang on Linux
With these changes, it is possible to build Nanos with the clang
compiler (by typing e.g. `make CC=clang`), as an alternative to
gcc.
The changes in klib/sandbox.h and src/x86_64/breakpoint.c fix the
"passing arguments to a function without a prototype is deprecated
in all versions of C and is not supported in C2x" compiler error.
libresolv.so.2 is being added to the shared libraries for the
runtime tests because when the webg Go program is built with
CC=clang it is dynamically linked to this library (in addition to
libc.so.6).
The OBJS_CRT* Make variables in rules.mk are being removed because
they are unused.
The other change in rules.mk is needed in order for clang to use a
global variable for the stack protector (upstream clang by default
uses TLS for the stack protector, whereas Apple's clang uses a
global variable).
The change in src/x86_64/gdb_machine.h fixes the "comparison of
unsigned expression >= 0 is always true" compiler error.
The changes in the acpica patch fix some "variable set but not
used" compiler errors.

Tested with clang version 15.0.7.