net/direct.c: fix tcp_pcb refcounting for client connections
When direct_connect() is called to establish a TCP client
connection, a new tcp_pcb structure is created and its refcount is
incremented. After the connection is established, a direct_conn
structure is created, a pointer to the tcp_pcb structure is stored
in direct_conn, and the tcp_pcb refcount is incremented again.
Then, when the connection is shut down, the refcount is decremented
in direct_conn_closed(), but not in direct_dealloc() (because the
tcp_pcb pointer in the direct structure is reset to 0 in order to
avoid calling tcp_close() on an already closed tcp_pcb). This
prevents the tcp_pcb from being deallocated, thereby creating a
memory leak.
This commit modifies direct_conn_alloc() so that the refcount is
not incremented after a TCP client connection is established
(i.e. the ownership of the tcp_pcb reference moves from the direct
structure to the direct_conn structure). This fixes the memory
leak.