poll(): fix events returned by filesystem entry file descriptors
Poll events returned by file descriptors corresponding to regular
files and directories are always (EPOLLIN | EPOLLOUT), regardless
of the current offset in the file descriptor. Polling on a symlink
file descriptor returns EPOLLNVAL.
This commit adds handling of POLLNVAL in the generic Unix poll code
and fixes the file_events closure to follow the above rules.