lseek: fix seeking on directory file descriptors
Setting the offset to the "end of the file" on a directory file
descriptor has the effect that no directory entries will be
returned at the next getdents syscall invocations. Setting the
offset past the "end of the file" should not be allowed.
This commit fixes the lseek() implementation to follow the above
rules.