Unix file struct: remove length field
Using the length field of the file struct to determine the length
of an open file can lead to incorrect results if the same file is
opened via multiple file descriptors at the same time, because a
change of file length resulting from an operation on a file
descriptor will not be reflected in the length field of the other
descriptors. This can cause problems such as for example the
inability to read file contents that have been written via a
different file descriptor (#1911).
This commit removes the length field from the file struct, and
changes the code so that fsfile_get_length() is used to determine
the length of an open file. The file length check in the
check_file_read() function is being removed because it is racy with
respect to concurrent file modifications, and the pageache code
can handle gracefully read requests past the end of a file (the
readv runtime test is being amended to exercise this behavior).