encode integer value sign separately from magnitude
This change alters the way signed numbers are encoded by using the length
field of the encoding header to store the integer sign and storing the number
magnitude in the varint. This avoids the need to encode negative numbers as
two's complement, sign-extended 64-bit integers, and also allows the full
range of unsigned 64-bit integers (0 to U64_MAX) to be represented by integer
values (albeit indirectly if the integer exceeds IMM_UINT_MAX). Printing of
integer values (by print_value_internal) has been corrected to print such
large unsigned integers correctly and not assume that all values are signed.