add non-creating fsfile_open(), fix behavior on missing program file
This adds fsfile_open(), which simply opens an existing file and returns its
fsfile on success. It returns null if the file does not exist or couldn't
otherwise be opened.

This also replaces the cryptic "cannot set on unknown value" error message
reported when the program binary specified in the manifest cannot be found.