demand_anonymous_page(): call storage_sync() via asynchronous thunk
A call to storage_sync() may suspend the current context, e.g. to
acquire the filesystem mutex. Therefore, this function cannot be
called directly from the code that handles a page fault exception;
instead, an asynchronous thunk should be used.
This change modifies the mmap_anon_page() closure function so that
it calls storage_sync() first, and the demand_anonymous_page()
function so that it invokes the above closure via async_apply_bh().

Closes #1933.