x86_64 pci_platform_init_bar(): fix handling of 64-bit addresses
The PCI BAR initialization code reads the base address configured
in a given BAR, and assigns a newly allocated address if the read
address is zero. The existing code is not handling 64-bit
addresses, thus fails to properly read and assign addresses when
the 64-bit flag is set.
This change fixes the above issue, and allows the virtio SCSI
driver to work correctly when running on an x86_64 machine with
OVMF UEFI firmware.