epoll: add EPOLLEXCLUSIVE support
This change adds support for the EPOLLEXCLUSIVE flag in the epoll
implementation, which allows limiting the number of threads waiting
on a file descriptor that are woken up when an event set is
notified. The EPOLLEXCLUSIVE flag works on a per file descriptor
basis, not on a per epoll instance basis, so it can be used with
threads waiting on a given file descriptor via one or more epoll
instances.
The event_handler closure type in the notify subsystem has been
changed to return a u64 value instead of a boolean value, so the
event handlers can indicate whether they have "consumed" the events
being reported. A notify entry is now associated to a set of flags,
used to indicate whether a handler is exclusive. When an event set
is notified, the notify subsystem calls all non-exclusive handlers,
plus it calls the exclusive handlers until the event set is
"consumed" (by any previously invoked handler, exclusive or
non-exclusive).
The notify subsytem is agnostic of the logic to determine whether
an event set is consumed; in the epoll implementation, an event
set is consumed when one or more waiting threads are unblocked. In
addition, the epoll_wait() implementation has been enhanced so that
a single event notification can wake up multiple threads blocked on
the same epoll instance.