Unix mprotect(): fix update of process virtual memory map
In the current code, the vmap_update_protections_intersection()
function is overwriting the entire contents of a vmap entry while
this entry is inserted in the process memory map; this can corrupt
the entry fields specific to the rangemap implementation (i.e. the
red-black tree), which can cause a crash when the map is accessed
later.
This change fixes the above issue by adding a vmap_set_offsets()
static function that overwrites only the vmap-specific fields that
need to be overwritten, and calling this function to overwrite the
contents of a vmap entry inserted in the virtual memory map.
altered_vmap_key() has been refactored to use this function.
This fixes a kernel hang observed when the mprotect() syscall is
invoked on a kernel built with the Clang compiler.