VGA console driver: check whether probed device is in VGA mode 3
A VM may have VGA-compatible devices (for example, GPUs) that are
not meant to be used as VGA displays, and the VGA console driver
should not bind to these devices, so that other drivers (e.g. GPU
drivers) can bind to them.
This change adds a check to the VGA driver probe function to verify
whether the device being probed has been set up in VGA mode 3
(which is the only mode supported by the VGA console driver).