Syscalls: add support for preadv and pwritev
These syscalls are analogous to readv and writev, but take an
additional file offset argument and do not change the internal file
offset.
The existing readv and writev functions have been refactored to de-
duplicate code, with a new iov_internal() function which is now
called by readv, writev, preadv and pwritev.
The readv and writev runtime tests have been amended to exercise
preadv and pwritev as well.