signalfd: fix polling for blocked signals
Commit fbfdb67eee5b8f1768d4fa9dacdedd49ab0ddd8b changed the
signalfd code so that signals retrieved via a read() are associated
to the thread that issues the read request (instead of the thread
that created the signalfd); this introduced a regression that
causes delivery of blocked signals to not be detected by a thread
that polls on a signalfd without reading from the signalfd.
This change fixes the above regression by moving the signalfds
notify set from the thread struct to the process struct, and adding
a signalfd to this notify set when the signalfd is created (instead
of when a read operation is requested). This allows the set of
signals of interest (from which the effective signal mask is
calculated) to be properly updated without the need for a user
thread to issue read operations on a signalfd.