Signals: fix handling of ignored signals
Signals being ignored by a user process should not interrupt
syscalls such as pause() and sigsuspend(), but should still be
detectable via sigwaitinfo(), sigtimedwait() and sigwait(), as well
as via signalfd. With the existing code, ignored signals are not
detectable by any means.
This change fixes the above issue by handling ignored signals
similarly to non-ignored signals, i.e. enqueing them in the
(process or thread) signal queue, and properly handling the
corresponding signal disposition when they are dequeued. The
`ignored` signal set is no longer used and has been removed from
the sigstate struct; the check for syscalls to be restarted has
been changed to differentiate between syscalls that should be
restarted based on the SA_RESTART flag, and syscalls that should be
restarted if no signal handler is installed (i.e. if an ignored
signal has been raised).
The signal runtime test has been amended to check the behavior of
sigwait() and signalfd with ignored signals.